Email Sumarizer/requirements.txt
google-api-python-client==2.142.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
requests==2.32.3
tiktoken==0.7.0
